pyproject.toml
[project]
name = "imtixon"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django[argon2]>=5.2.7",
    "django-jazzmin>=3.0.1",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "djangorestframework-simplejwt>=5.5.1",
    "booktype>=1.5",
    "redis==5.0.1",
]
